backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.2.2
openpyxl==3.1.5
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
pypdf==3.17.4
markdown==3.5.1
chromadb==0.4.18
python-multipart==0.0.6
langchain==0.0.350
langchain-community==0.0.10
google-generativeai==0.4.1
langchain-google-genai==0.0.11
rapidfuzz>=3.0.0
